+++ src/theme/variables/testDark.ts
@@ -384,6 +384,25 @@ export const testDark: Theme = {
   outlineTextInputLabelColorDisabled: palette.gray,
   outlineTextInputLabelColorFocused: palette.black,
 
+  // Simple Text Input
+  textInputTextColor: palette.gray,
+  textInputTextColorDisabled: palette.white,
+  textInputTextColorFocused: palette.white,
+  textInputBackgroundColor: palette.edgeNavy,
+  textInputBackgroundColorDisabled: palette.edgeNavy,
+  textInputBackgroundColorFocused: palette.edgeNavy,
+  textInputBorderColor: palette.mutedBlue,
+  textInputBorderColorDisabled: palette.gray,
+  textInputBorderColorFocused: palette.edgeMint,
+  textInputBorderWidth: 1,
+  textInputBorderRadius: 100,
+  textInputIconColor: palette.mutedBlue,
+  textInputIconColorDisabled: palette.gray,
+  textInputIconColorFocused: palette.edgeMint,
+  textInputPlaceholderColor: palette.gray,
+  textInputPlaceholderColorDisabled: palette.gray,
+  textInputPlaceholderColorFocused: palette.edgeMint,
+
   // Animation
   fadeDisable: palette.gray,