+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - added: Automatically enable tokens with balances
 - added: Add Pokt RPCs as option for Fantom
 - added: GLIF Infinity Pool staking plugin for Filecoin FEVM currency
+- added: New SimpleTextInput component
 - added: Support for Thorchain Savers EVM token deposits
 - added: Add Solana init options
 - added: BSC staking on Thorchain Savers